index

+++ SeramikStore.Web/Controllers/ProductController.cs
@@ -1,19 +1,25 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
+using Microsoft.EntityFrameworkCore;
 using SeramikStore.Entities;
 using SeramikStore.Services;
 using SeramikStore.Web.ViewModels;
+using System.Collections.Generic;
 using System.Globalization;
 
 public class ProductController : Controller
 {
     private readonly IProductService _productService;
     private readonly ICurrencyService _currencyService;
+    private readonly IProductImageService _productImageService;
+    private readonly ICategoryService _categoryService;
 
-    public ProductController(IProductService productService, ICurrencyService currencyService)
+    public ProductController(IProductService productService, ICurrencyService currencyService, IProductImageService productImageService, ICategoryService categoryService)
     {
         _productService = productService;
         _currencyService = currencyService;
+        _productImageService = productImageService;
+        _categoryService = categoryService;
     }
 
     public IActionResult Index()
@@ -27,12 +33,19 @@ public class ProductController : Controller
     {
         var model = new ProductCreateViewModel
         {
-           // Product = new Product(),
+            // Product = new Product(),
             Currencies = _currencyService.CurrencyList()
                         .Select(x => new SelectListItem
                         {
                             Value = x.Id.ToString(),
                             Text = $"{x.Code} - {x.Name}"
+                        }).ToList(),
+
+            Categories = _categoryService.CategoryList()
+                        .Select(x => new SelectListItem
+                        {
+                            Value = x.Id.ToString(),
+                            Text = x.Name.ToString()
                         }).ToList(),
         };
 
@@ -52,6 +65,13 @@ public class ProductController : Controller
                     Text = $"{x.Code} - {x.Name}"
                 }).ToList();
 
+            model.Categories = _categoryService.CategoryList()
+            .Select(x => new SelectListItem
+            {
+                Value = x.Id.ToString(),
+                Text = x.Name.ToString()
+            }).ToList();
+
             return View(model);
         }
 
@@ -60,6 +80,8 @@ public class ProductController : Controller
             model.UnitPrice.Replace(".", "").Replace(",", "."),
             CultureInfo.InvariantCulture);
 
+
+
         var product = new Product
         {
             ProductCode = model.ProductCode,
@@ -67,6 +89,7 @@ public class ProductController : Controller
             ProductDesc = model.ProductDesc,
             UnitPrice = unitPrice,
             CurrencyId = model.CurrencyId,
+            CategoryId = model.CategoryId,
             AvailableForSale = model.AvailableForSale
         };
 
@@ -79,20 +102,88 @@ public class ProductController : Controller
     [HttpGet]
     public IActionResult Edit(int id)
     {
-        return View(_productService.ProductGetById(id));
+        var product = _productService.ProductGetById(id);
+
+        if (product == null)
+        {
+            return NotFound(); // veya RedirectToAction("Index")
+        }
+
+        var model = new ProductEditViewModel
+        {
+            Id = product.Id,
+            ProductCode = product.ProductCode,
+            ProductName = product.ProductName,
+            ProductDesc = product.ProductDesc,
+            UnitPrice = product.UnitPrice.ToString("N2", new CultureInfo("tr-TR")),
+            CurrencyId = product.CurrencyId,
+            CategoryId = product.CategoryId,
+            AvailableForSale = product.AvailableForSale,
+
+            Currencies = _currencyService.CurrencyList()
+                .Select(x => new SelectListItem
+                {
+                    Value = x.Id.ToString(),
+                    Text = $"{x.Code} - {x.Name}"
+                })
+                .ToList(),
+
+            Categories = _categoryService.CategoryList()
+                        .Select(x => new SelectListItem
+                        {
+                            Value = x.Id.ToString(),
+                            Text = x.Name.ToString()
+                        }).ToList(),
+        };
+
+        return View(model);
     }
 
     [HttpPost]
     [ValidateAntiForgeryToken]
-    public IActionResult Edit(Product product)
+    public IActionResult Edit(ProductEditViewModel model)
     {
         if (!ModelState.IsValid)
-            return View(product);
+        {
+            model.Currencies = _currencyService.CurrencyList()
+                .Select(x => new SelectListItem
+                {
+                    Value = x.Id.ToString(),
+                    Text = $"{x.Code} - {x.Name}"
+                }).ToList();
+
+            model.Categories = _categoryService.CategoryList()
+            .Select(x => new SelectListItem
+            {
+                Value = x.Id.ToString(),
+                Text = x.Name.ToString()
+            }).ToList();
+
+            return View(model);
+        }
+
+        var unitPrice = decimal.Parse(
+            model.UnitPrice.Replace(".", "").Replace(",", "."),
+            CultureInfo.InvariantCulture
+        );
+
+        var product = new Product
+        {
+            Id = model.Id,
+            ProductCode = model.ProductCode,
+            ProductName = model.ProductName,
+            ProductDesc = model.ProductDesc,
+            UnitPrice = unitPrice,
+            CurrencyId = model.CurrencyId,
+            CategoryId = model.CategoryId,
+            AvailableForSale = model.AvailableForSale
+        };
 
         _productService.UpdateProduct(product);
-        TempData["Success"] = "Ürün güncellendi";
-        return RedirectToAction(nameof(Index));
+
+        return RedirectToAction("Index");
     }
+
 
     [HttpGet]
     public IActionResult Delete(int id)
@@ -108,8 +199,24 @@ public class ProductController : Controller
     [ValidateAntiForgeryToken]
     public IActionResult DeleteConfirmed(int id)
     {
-        _productService.DeleteProduct(id);
-        TempData["Success"] = "Product deleted successfully.";
-        return RedirectToAction(nameof(Index));
+
+
+        List<ProductImage> images = _productImageService.GetByProductId(id);
+
+        if (images.Any())
+        {
+            ModelState.AddModelError("", "Bu ürüne ait resim bulunduğu için silinemez.");
+
+            var product = _productService.ProductGetById(id);
+            return View("Delete", product);
+        }
+        else
+        {
+            _productService.DeleteProduct(id);
+            TempData["Success"] = "Product deleted successfully.";
+            return RedirectToAction(nameof(Index));
+        }
+
+
     }
 }
